[Feat] 플레이어의 스킬 사용 구현과 사망시 게임종료
플레이어의 스킬 사용 구현과 사망시 게임종료

+++ TextRPGTeam30/OffensiveSkill.cs
@@ -14,7 +14,7 @@
             this.damageModifier = damageModifier;
         }
 
-        public int UseSkill(int damage)
+        public float UseSkill(float damage)
         {
             base.UseSkill();
